LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/animation-container-type-dynamic-expected.txt

Test

FAIL Animated container creating new container assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 255)"
